+++ tests/test_main.py
@@ -7,9 +7,9 @@ from src.main import main
 
 class TestProvideStatistics(TestCase):
 
-    @patch("send_statistics.Publisher.send_statistics", return_value=True)
-    @patch("get_statistics.Collector.get_statistics", return_value='{"Foo": "Bar"}')
-    @patch("check_container.Container.is_running", return_value=True)
+    @patch("src.send_statistics.Publisher.send_statistics", return_value=True)
+    @patch("src.get_statistics.Collector.get_statistics", return_value='{"Foo": "Bar"}')
+    @patch("src.check_container.Container.is_running", return_value=True)
     def test_happy_flow(self, cntnr_running, stats, msg_sent):
         with self.assertLogs("__main__", level="DEBUG") as log:
             logging.getLogger("__main__").debug("Unbound statistics sent successful.")
@@ -18,7 +18,7 @@ class TestProvideStatistics(TestCase):
         self.assertEqual(result, None)
         self.assertEqual(log.output, ["DEBUG:__main__:Unbound statistics sent successful."])
 
-    @patch("check_container.Container.is_running", return_value=False)
+    @patch("src.check_container.Container.is_running", return_value=False)
     def test_failed(self, cntnr_running):
         with self.assertLogs("__main__", level="ERROR") as log:
             logging.getLogger("__main__").error("A Container is missing or not running. Check log for details.")
