+++ tests/test_send_statistics.py
@@ -9,7 +9,7 @@ from src.send_statistics import Publisher
 
 class TestSendStatistics(TestCase):
 
-    @patch("send_statistics.docker.execute", side_effect=DockerException(return_code=1,
+    @patch("src.send_statistics.docker.execute", side_effect=DockerException(return_code=1,
                                                                          command_launched="/usr/bin/docker exec publisher_cntnr pub -h 127.0.0.1 -q 1 -t statistics/foo/bar -m {'Foo':'Bar'}",
                                                                          stderr=b'f\x00o\x00o\x00b\x00a\x00r\x00b\x00a\x00z\x00'))
     def test_send_msg_failed(self, error):
@@ -18,7 +18,7 @@ class TestSendStatistics(TestCase):
         self.assertRaises(DockerException)
         self.assertFalse(success)
 
-    @patch("send_statistics.docker.execute", return_value=True)
+    @patch("src.send_statistics.docker.execute", return_value=True)
     def test_send_msg_successful(self, result):
         stats = {"thread0.num.queries": 1102}
         self.assertTrue(Publisher.send_statistics("127.0.0.1", stats, "publisher_cntnr"))
